Style p below legend element
You want to tell user that all of the fields that have
star symbol is required.

also you want to style the bg color of a required inputs.
star symbol should be colored red.

you want to tell user that they have filled the field properly with
outline colored to green indicating that it's a success.